Fix permission denied error when creating staging dir
Avoid using the ansible_user_dir when creating the staging directory for
the buildbot worker configuration.  This variable is not reliable since
sometimes ansible_user_dir will have the value of the root user, rather
than the expected home directory of the ansible_ssh_user.

Instead, just use become to also setup the staging directory and use the
constant path.