Fix buildbot-worker install on CentOS-7
Recent changes introduced the requirement for rust, so we need to
bootstrap by first creating the virtualenv, then upgrading pip in the
virtualenv, and finally installing the buildbot-worker package with the
upgraded pip.

This dance seems difficult to do with the pip module, so just fallback
to a series of commands with changed args for idempotency.