Update dependencies for Solaris buildbot-workers
Update the dependencies need to install a current version of
the buildbot-worker package on Solaris.

The Python cryptography package now requires a rust compiler (since
binaries are not distributed for Solaris).

The Python cryptography package now requires openssl version 1.1.
Solaris has both versions 1.0 and 1.1, and the IPS mediator can be used
to select 1.1 as the current version for builds and run-time.

The ffi extension requires the 'pkg-config' tool and an env var to
specify 64 bit builds.

Update the python3 version installed, which is now python 3.7 on
Solaris 11.4.x

Remove the check for `cc`, since the rustc installs a version of gcc
anyway.

Create the python virtualenv first and install the latest pip before
installing buildbot-worker and it's dependencies.