worker: Remove buildbot worker info file workaround
Get rid of the "staging" work-around which was used to update the
buildbot worker info files (host, admin, access_uri).  This was done to
work around the tmp directory permission warnings when the ansible user
was different than the non-root user that owned the destination files.

The workaround was was awful, ugly, error prone, and completely
unnecessary.  As it turns out, all this trouble was due to the fact that
the "copy" module creates a tmp file on the remote when the source is
"content" in order to render the content.

Instead, just add three small, jinja2 template files, and use the
template module to update the info files.  The content is rendered on
the controller and is put into place with the correct owner and
permissions.