worker: Support python 2.7 on CentOS-6
CentOS 6 shipped with python 2.6 which is too old for the buildbot.  We
currently support installing a buildbot worker with python3.6 installed
as a RH software collection.

In addition, support python 2.7 which was installed from source on the
target machine before running Ansible.

The new `buildbot_worker_python` role variable must be set to the path
of the python used to install and run the buildbot worker, for example
buildbot_worker_python=/usr/local/bin/python2.7.