master: Add buildbotctl rotatelog
Instead of always rotating the log on restart, add a separate
buildbotctl command to rotate the twistd.logs.

Rewrite the rotatelog shell function to actually rotate the logs.  The
name of the twistd.log was incorrect, so the logs on the master were not
actually rotated.

Copy the contents from twistd.log to twistd.log.1 then truncate the
twistd.log since the file descriptor can be open by the buildbot master.
This lets us rotate the log without stopping the buildbot process.