worker: Detect the user group on non-macos platforms again
The buildbot user account creation task was changed for the macOS
platform to work around what looks to be a bug. On that platform we need
to specify "staff" as the user's primary group in the version of Ansible
that is currently in use.

On linux the primary group is "buildbot", and that group is created
the first time the user task is run, unless we specify the group name,
in which case the task assumes the group already exists. And since
it doesn't the user task fails.

So go back to detecting the group name after the task on all the non
macos platfoms like we did before, and on macos, specify "staff" which
is known to exisit.